pyproject.toml
[project]
name = "task-management-api"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi[standard]>=0.128.0",
    "psycopg2-binary>=2.9.11",
    "pydantic-settings>=2.12.0",
    "pytest>=9.0.2",
    "sqlmodel>=0.0.31",
    "pwdlib>=0.1.1",
    "pydantic[email]>=2.12.5",
    "python-jose[cryptography]>=3.5.0",
]
